+++ src/main/java/com/telran/libraryapp/entity/Author.java
@@ -15,9 +15,10 @@ import java.util.Objects;
 @NoArgsConstructor
 @Entity
 public class Author {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String name;
     private String surname;
     private String authorInfo;
